requirements.txt
requests
tabulate
colorama
matplotlib
pytz
python-dotenv